separate out repo and chart-related functionality and tests for easier readability and add unit tests for GetAvailablePackageDetail
